Revert "Feat/plug logic ui"
