+++ TicTacToe/AppDelegate.swift
@@ -16,7 +16,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         
-        let gameViewController = GameViewController(game: Game())
+        let gameViewController = GameViewController()
         navigationController = UINavigationController(rootViewController: gameViewController)
         
         window = UIWindow(frame: UIScreen.main.bounds)
