--- TicTacToe/Extensions/UIStackView+Extensions.swift
@@ -1,18 +1,0 @@
-//
-//  UIStackView+Extensions.swift
-//  TicTacToe
-//
-//  Created by Ananth Bhamidipati on 20/06/2020.
-//  Copyright © 2020 2020-DEV-075. All rights reserved.
-//
-
-import UIKit
-
-extension UIStackView {
-    func addBackground(color: UIColor) {
-        let subView = UIView(frame: bounds)
-        subView.backgroundColor = color
-        subView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
-        insertSubview(subView, at: 0)
-    }
-}
