--- TicTacToe/Game/GameStatusInfo.swift
@@ -1,18 +1,0 @@
-//
-//  GameStatusInfo.swift
-//  TicTacToe
-//
-//  Created by Ananth Bhamidipati on 21/06/2020.
-//  Copyright © 2020 2020-DEV-075. All rights reserved.
-//
-
-import Foundation
-
-enum GameStatusInfo {
-    case won
-    case draw
-    case alreadyPlayed
-    case markMove
-    case nextMove
-    case gameOver
-}
