+++ TicTacToe/Game/GameViewController.swift
@@ -1,0 +1,38 @@
+//
+//  GameViewController.swift
+//  TicTacToe
+//
+//  Created by Ananth Bhamidipati on 20/06/2020.
+//  Copyright © 2020 2020-DEV-075. All rights reserved.
+//
+
+import UIKit
+
+class GameViewController: UIViewController {
+    
+    init() {
+        super.init(nibName: nil, bundle: nil)
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
