--- TicTacToe/Game/Player.swift
@@ -1,14 +1,0 @@
-//
-//  Player.swift
-//  TicTacToe
-//
-//  Created by Ananth Bhamidipati on 21/06/2020.
-//  Copyright © 2020 2020-DEV-075. All rights reserved.
-//
-
-import Foundation
-
-enum Player {
-    case playerX
-    case playerO
-}
